+++ src/main/java/com/beanions/common/controller/MainController.java
@@ -30,10 +30,4 @@ public class MainController {
         return "user/login";
     }
 
-    // feature/adminPost
-    @GetMapping("/admin")
-    public String admin() {
-        return "admin/main/dashBoard";
-    }
-
 }
