+++ src/main/resources/mapper/AdminMapper.xml
@@ -272,4 +272,57 @@
          ORDER BY POST_DATE DESC
          LIMIT 5
     </select>
+    <select id="selectCurrentMagazine" resultMap="selectCurrenBoardResultMap">
+        SELECT
+               POST_CODE
+             , MEMBER_CODE
+             , POST_TITLE
+             , POST_CONTEXT
+             , POST_DATE
+             , MAIN_CATEGORY
+             , SUB_CATEGORY
+             , REVIEW_STATUS
+             , VIEW_COUNT
+          FROM POST
+         WHERE MAIN_CATEGORY = '매거진'
+         ORDER BY POST_DATE DESC
+         LIMIT 5
+    </select>
+
+    <resultMap id="selectAdminMainResultMap" type="com.beanions.common.dto.AdminMainDTO">
+        <result property="memberAll" column="member_all"/>
+        <result property="memberDay" column="member_day"/>
+        <result property="postAll" column="post_all"/>
+        <result property="postDay" column="post_day"/>
+        <result property="reviewAll" column="review_all"/>
+        <result property="reviewDay" column="review_day"/>
+        <result property="yesinAll" column="yesin_all"/>
+        <result property="yesinDay" column="yesin_day"/>
+        <result property="yerangAll" column="yerang_all"/>
+        <result property="yerangDay" column="yerang_day"/>
+    </resultMap>
+    <select id="selectAdminMainData" resultMap="selectAdminMainResultMap">
+        SELECT
+        A.MEMBER_ALL
+        , B.MEMBER_DAY
+        , C.POST_ALL
+        , D.POST_DAY
+        , E.REVIEW_ALL
+        , F.REVIEW_DAY
+        , G.YESIN_ALL
+        , H.YESIN_DAY
+        , I.YERANG_ALL
+        , J.YERANG_DAY
+        FROM
+        (SELECT COUNT(*) MEMBER_ALL FROM MEMBERS) A
+        , (SELECT COUNT(*) MEMBER_DAY FROM MEMBERS WHERE SIGNUP_DATE = CURDATE()) B
+        , (SELECT COUNT(*) POST_ALL FROM POST) C
+        , (SELECT COUNT(*) POST_DAY FROM POST WHERE POST_DATE = CURDATE()) D
+        , (SELECT COUNT(*) REVIEW_ALL FROM POST WHERE MAIN_CATEGORY = '리뷰') E
+        , (SELECT COUNT(*) REVIEW_DAY FROM POST WHERE MAIN_CATEGORY = '리뷰' AND POST_DATE = CURDATE()) F
+        , (SELECT COUNT(*) YESIN_ALL FROM POST WHERE SUB_CATEGORY = '예신') G
+        , (SELECT COUNT(*) YESIN_DAY FROM POST WHERE SUB_CATEGORY = '예신' AND POST_DATE = CURDATE()) H
+        , (SELECT COUNT(*) YERANG_ALL FROM POST WHERE SUB_CATEGORY = '예랑') I
+        , (SELECT COUNT(*) YERANG_DAY FROM POST WHERE SUB_CATEGORY = '예랑' AND POST_DATE = CURDATE()) J;
+    </select>
 </mapper>
