+++ src/main/resources/templates/admin/main/dashBoard.html
@@ -14,7 +14,7 @@
 <article>
 <!--    <h1 align="center">관리자 메인</h1>-->
     <p class="dashBoard_member_text">회원</p>
-    <div class="dashBoard_box_first">
+    <div class="dashBoard_box_first" th:each="data : ${ aminMainData }">
         <div class="dashBoard_item_first">
             <div class="main_content_box">
                 <p class="content_header">방문자 수</p>
@@ -37,11 +37,11 @@
                 <div class="content_text_box">
                     <div class="content_count_box">
                         <p>전체</p>
-                        <p class="number">200</p>
+                        <p class="number" th:text="${data.memberAll}"></p>
                     </div>
                     <div class="content_count_box">
                         <p>오늘</p>
-                        <p class="number">2</p>
+                        <p class="number" th:text="${data.memberDay}"></p>
                     </div>
                 </div>
                 <hr>
@@ -53,18 +53,18 @@
                 <div class="content_text_box">
                     <div class="content_count_box">
                         <p>전체</p>
-                        <p class="number">80</p>
+                        <p class="number" th:text="${ data.postAll }"></p>
                     </div>
                     <div class="content_count_box">
                         <p>오늘</p>
-                        <p class="number">2</p>
+                        <p class="number" th:text="${ data.postDay }"></p>
                     </div>
                 </div>
                 <hr>
             </div>
         </div>
     </div>
-    <div class="dashBoard_box_second">
+    <div class="dashBoard_box_second" th:each="data : ${ aminMainData }">
         <div class="dashBoard_board_box">
             <p class="dashBoard_board_text">예리뷰</p>
             <div class="dashBoard_item_second">
@@ -73,11 +73,11 @@
                     <div class="content_text_box">
                         <div class="content_count_box">
                             <p>전체</p>
-                            <p class="number">20</p>
+                            <p class="number" th:text="${ data.reviewAll }"></p>
                         </div>
                         <div class="content_count_box">
                             <p>오늘</p>
-                            <p class="number">1</p>
+                            <p class="number" th:text="${ data.reviewDay}"></p>
                         </div>
                     </div>
                     <hr>
@@ -93,21 +93,21 @@
                         <div class="free_count_box" id="yesin_count_box">
                             <div class="free_content_count_box">
                                 <p>예신</p>
-                                <p class="number">40</p>
+                                <p class="number" th:text="${ data.yesinAll }"></p>
                             </div>
                             <div class="free_content_count_box" id="yesin_margin">
                                 <p>신규</p>
-                                <p class="number">4</p>
+                                <p class="number" th:text="${ data.yesinDay }"></p>
                             </div>
                         </div>
                         <div  class="free_count_box">
                             <div class="free_content_count_box" id="yerang_margin">
                                 <p>예랑</p>
-                                <p class="number">40</p>
+                                <p class="number" th:text="${ data.yerangAll }"></p>
                             </div>
                             <div class="free_content_count_box">
                                 <p>신규</p>
-                                <p class="number">0</p>
+                                <p class="number" th:text="${ data.yerangDay }"></p>
                             </div>
                         </div>
                     </div>
@@ -128,11 +128,13 @@
                         <th class="dashBoard_th">작성일</th>
                         <th class="dashBoard_th">조회수</th>
                     </tr>
-                    <tr class="dashBoard_tr" >
-                        <td class="dashBoard_td" ></td>
-                        <td class="dashBoard_td" ></td>
-                        <td class="dashBoard_td" ></td>
-                        <td class="dashBoard_td" ></td>
+                    <tr class="dashBoard_tr" th:each="notice : ${ postList }">
+                        <td class="dashBoard_td" th:text="${ notice.postCode}"></td>
+                        <td class="dashBoard_td">
+                            <a th:text="${notice.postTitle}" th:href="@{/admin/notice/detail(id=${notice.postCode})}"></a>
+                        </td>
+                        <td class="dashBoard_td" th:text="${ #dates.format(notice.postDate, 'yyyy-MM-dd') }"></td>
+                        <td class="dashBoard_td" th:text="${ notice.viewCount }"></td>
                     </tr>
                 </table>
             </div>
@@ -147,11 +149,13 @@
                         <th class="dashBoard_th">작성일</th>
                         <th class="dashBoard_th">조회수</th>
                     </tr>
-                    <tr class="dashBoard_tr">
-                        <td class="dashBoard_td">37</td>
-                        <td class="dashBoard_td">홈페이지 가이드라인 신규 회원 필독!!!</td>
-                        <td class="dashBoard_td"></td>
-                        <td class="dashBoard_td"></td>
+                    <tr class="dashBoard_tr" th:each="magazine : ${ magazineList }">
+                        <td class="dashBoard_td" th:text="${ magazine.postCode }"></td>
+                        <td class="dashBoard_td">
+                            <a th:text="${magazine.postTitle}" th:href="@{/admin/magazine/detail(id=${magazine.postCode})}"></a>
+                        </td>
+                        <td class="dashBoard_td" th:text="${ #dates.format(magazine.postDate, 'yyyy-MM-dd')}"></td>
+                        <td class="dashBoard_td" th:text="${ magazine.viewCount }"></td>
                     </tr>
                 </table>
             </div>
